Feature user history

+++ restful-web-services/src/main/java/com/sept/rest/webservices/restfulwebservices/post/PostController.java
@@ -23,8 +23,10 @@ import org.springframework.stereotype.Repository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+
 import com.sept.rest.webservices.restfulwebservices.jwt.JwtTokenUtil;
 import com.sept.rest.webservices.restfulwebservices.post.Post.Status;
+import com.sept.rest.webservices.restfulwebservices.user.UserService;
 
 
 
@@ -36,10 +38,14 @@ public class PostController {
 
 	@Autowired
     private JwtTokenUtil jwtTokenUtil;
+	
+	@Autowired
+    private UserService checker;
 
 	@Autowired
 	private PostService service;
 	
+
 
 	// posts that are for sale
 	@GetMapping("/posts")
@@ -63,8 +69,7 @@ public class PostController {
 	@PostMapping("/postitem")
 	public Post addPost(@RequestBody Post post, HttpServletRequest request)
 	{
-		System.out.println("adding post: " + post);
-		post.setOwner(getOwnerId(request));
+		post.setOwner(checker.getOwnerId(request));
 		return service.update(post);
 	}
 
@@ -72,14 +77,17 @@ public class PostController {
 	@PutMapping("/posts/{id}")
 	public Post updatePost(@PathVariable Long id, @RequestBody Post edit, HttpServletRequest request)
 	{
-		System.out.println("received post backend: " + edit);
-		return correctOwner(id, request) ? service.update(edit) : null;
+		return checker.isEqual(edit.getOwnerId(), request) ? service.update(edit) : null;
 	}
+
 	
 	@DeleteMapping("/posts/{id}")
 	public void deletePost(@PathVariable Long id, HttpServletRequest request)
 	{	
-		service.delete(service.get(id));
+		Post toDelete = service.get(id);
+		if (checker.isEqual(toDelete.getOwnerId(), request)) {
+			service.delete(service.get(id));
+		}
 	}
 	
 	@GetMapping("/posts/{id}")
@@ -87,10 +95,7 @@ public class PostController {
 	{
 		return service.get(id);
 	}
-
 	
-	
-
 	@PostMapping("/posts/{id}/{status}")
 	public Post updatePostStatus(@PathVariable Long id, @PathVariable Status status, HttpServletRequest request)
 	{	
@@ -110,20 +115,64 @@ public class PostController {
 	
 	
 	
-	private String getOwnerId(HttpServletRequest request)
-	{
-		return jwtTokenUtil.getUsernameFromToken(request.getHeader("Authorization").substring(7));
-	}
+//	reused /posts/{id}/{status} (already done for sold, available)
+//	@PostMapping("/postsdelete")
+//	public void deletePosts(@RequestBody Post post, HttpServletRequest request)
+//	{	
+//		if (correctOwner(post.getId(), request))
+//			post.setStatus(Status.DELETED);
+//	}
+//	
+// 	
+//	@PostMapping("/postspermdelete")
+//	public void DeletePosts(@RequestBody Post post, HttpServletRequest request)
+//	{	
+//		if (correctOwner(post.getId(), request))
+//			service.delete(service.get(post.getId()));
+//	}
 	
-	private boolean correctOwner(Long id, HttpServletRequest request)
+
+
+	
+	//logic for account history incase u need to sort them into groups like all marked as sold, all marked as deleted etc
+	// directly putting status, as you need to mention this in the url anyway
+	
+	@GetMapping("/{ownerId}/posts/{status}")
+	public List<Post> getCurrentPosts(@PathVariable String ownerId, @PathVariable Status status) 
 	{
-		Post realPost = service.get(id);
-		String loggedOwner =getOwnerId(request);
-		
-		if (!realPost.getOwnerId().equals(loggedOwner))
-			throw new NullPointerException("Error Update: Edit User ID " + loggedOwner + " does not match Item Owner ID " + realPost.getOwnerId());
-		return true;
+		return service.getOwnerPosts(ownerId, status);
 	}
+			
+		
+	
+
+
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+//	private String getOwnerId(HttpServletRequest request)
+//	{
+//		return jwtTokenUtil.getUsernameFromToken(request.getHeader("Authorization").substring(7));
+//	}
+//	
+//	private boolean correctOwner(Long id, HttpServletRequest request)
+//	{
+//		Post realPost = service.get(id);
+//		String loggedOwner =getOwnerId(request);
+//		
+//		if (!realPost.getOwnerId().equals(loggedOwner))
+//			throw new NullPointerException("Error Update: Edit User ID " + loggedOwner + " does not match Item Owner ID " + realPost.getOwnerId());
+//		return true;
+//	}
